requirements.txt
streamlit
python-docx
qrcode